requirements.txt
u-msgpack-python
